recipe_lab/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.3
faiss-cpu==1.8.0.post1
numpy==1.26.4
pydantic==2.9.2
orjson==3.10.7
pypdf==5.1.0
